app/requirements.txt
asgiref==3.9.1
attrs==25.3.0
Django==4.2
django-rest-authtoken==2.1.4
djangorestframework==3.14.0
gunicorn==20.1.0
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pytest==7.2.0
pytest-django==4.5.0
pytz==2025.2
setuptools==80.9.0
sqlparse==0.5.3
